@@ -41,6 +41,7 @@
   "homepage": "https://github.com/angular/angular-cli",
   "dependencies": {
     "@ngtools/json-schema": "^1.1.0",
+    "@angular-devkit/build-optimizer": "0.0.3",
     "autoprefixer": "^6.5.3",
     "chalk": "^1.1.3",
     "circular-dependency-plugin": "^3.0.0",
@@ -54,21 +55,22 @@
     "ember-cli-string-utils": "^1.0.0",
     "enhanced-resolve": "^3.1.0",
     "exports-loader": "^0.6.3",
-    "extract-text-webpack-plugin": "^2.1.0",
+    "extract-text-webpack-plugin": "3.0.0",
     "file-loader": "^0.10.0",
     "fs-extra": "~3.0.1",
     "get-caller-file": "^1.0.0",
     "glob": "^7.0.3",
     "heimdalljs": "^0.2.4",
     "heimdalljs-logger": "^0.1.9",
-    "html-webpack-plugin": "^2.19.0",
+    "html-webpack-plugin": "^2.29.0",
     "inflection": "^1.7.0",
     "inquirer": "^3.0.0",
     "isbinaryfile": "^3.0.0",
     "istanbul-instrumenter-loader": "^2.0.0",
     "json-loader": "^0.5.4",
+    "karma-source-map-support": "^1.2.0",
     "less": "^2.7.2",
-    "less-loader": "^4.0.2",
+    "less-loader": "^4.0.5",
     "license-webpack-plugin": "^0.4.3",
     "loader-utils": "^1.0.2",
     "lodash": "^4.11.1",
@@ -84,7 +86,6 @@
     "pug-html-loader": "^1.1.4",
     "raw-loader": "^0.5.1",
     "resolve": "^1.1.7",
-    "rimraf": "^2.5.3",
     "rsvp": "^3.0.17",
     "rxjs": "^5.0.1",
     "sass-loader": "^6.0.3",
@@ -100,53 +101,49 @@
     "typescript": "~2.3.1",
     "url-loader": "^0.5.7",
     "walk-sync": "^0.3.1",
-    "webpack": "~2.4.0",
-    "webpack-dev-middleware": "^1.10.2",
-    "webpack-dev-server": "~2.4.5",
+    "webpack": "~3.3.0",
+    "webpack-dev-middleware": "^1.11.0",
+    "webpack-dev-server": "~2.5.1",
     "webpack-merge": "^2.4.0",
-    "zone.js": "^0.8.4"
+    "zone.js": "^0.8.14"
   },
   "devDependencies": {
     "@angular/compiler": "^4.0.0",
     "@angular/compiler-cli": "^4.0.0",
     "@angular/core": "^4.0.0",
-    "@types/chai": "^3.4.32",
     "@types/chalk": "^0.4.28",
     "@types/common-tags": "^1.2.4",
     "@types/denodeify": "^1.2.29",
     "@types/express": "^4.0.32",
     "@types/fs-extra": "~3.0.2",
     "@types/glob": "^5.0.29",
-    "@types/jasmine": "~2.2.0",
+    "@types/jasmine": "2.5.45",
     "@types/lodash": "4.14.50",
     "@types/minimist": "^1.2.0",
     "@types/mock-fs": "^3.6.30",
     "@types/node": "^6.0.36",
     "@types/request": "0.0.39",
-    "@types/rimraf": "0.0.28",
     "@types/semver": "^5.3.30",
     "@types/source-map": "^0.5.0",
     "@types/webpack": "^2.2.15",
-    "chai": "^3.5.0",
     "conventional-changelog": "^1.1.0",
     "dtsgenerator": "^0.9.1",
     "eslint": "^3.11.0",
     "express": "^4.14.0",
-    "jasmine": "^2.4.1",
-    "jasmine-spec-reporter": "^3.2.0",
+    "jasmine": "^2.6.0",
+    "jasmine-spec-reporter": "^4.1.0",
     "minimist": "^1.2.0",
-    "mocha": "^3.2.0",
     "mock-fs": "^4.0.0",
     "npm-run": "^4.1.0",
     "npm-run-all": "^4.0.0",
-    "object-assign": "^4.0.1",
     "request": "^2.74.0",
     "resolve-bin": "^0.4.0",
     "rewire": "^2.5.1",
     "spdx-satisfies": "^0.1.3",
+    "tar": "^3.1.5",
     "through": "^2.3.6",
     "tree-kill": "^1.0.0",
-    "ts-node": "^2.0.0",
+    "ts-node": "^3.0.6",
     "tslint": "^5.1.0"
   },
   "optionalDependencies": {
