@@ -1,6 +1,6 @@
 {
   "name": "@angular/cli",
-  "version": "1.2.0-rc.1",
+  "version": "1.3.0-beta.1",
   "description": "CLI tool for Angular",
   "main": "lib/cli/index.js",
   "trackingCode": "UA-8594346-19",
@@ -27,8 +27,9 @@
   },
   "homepage": "https://github.com/angular/angular-cli",
   "dependencies": {
+    "@angular-devkit/build-optimizer": "0.0.3",
     "@ngtools/json-schema": "1.1.0",
-    "@ngtools/webpack": "1.5.0-rc.1",
+    "@ngtools/webpack": "1.6.0-beta.1",
     "autoprefixer": "^6.5.3",
     "chalk": "^1.1.3",
     "circular-dependency-plugin": "^3.0.0",
@@ -41,20 +42,21 @@
     "ember-cli-normalize-entity-name": "^1.0.0",
     "ember-cli-string-utils": "^1.0.0",
     "exports-loader": "^0.6.3",
-    "extract-text-webpack-plugin": "^2.1.0",
+    "extract-text-webpack-plugin": "3.0.0",
     "file-loader": "^0.10.0",
     "fs-extra": "^3.0.1",
     "get-caller-file": "^1.0.0",
     "glob": "^7.0.3",
     "heimdalljs": "^0.2.4",
     "heimdalljs-logger": "^0.1.9",
-    "html-webpack-plugin": "^2.19.0",
+    "html-webpack-plugin": "^2.29.0",
     "inflection": "^1.7.0",
     "inquirer": "^3.0.0",
     "isbinaryfile": "^3.0.0",
     "json-loader": "^0.5.4",
+    "karma-source-map-support": "^1.2.0",
     "less": "^2.7.2",
-    "less-loader": "^4.0.2",
+    "less-loader": "^4.0.5",
     "license-webpack-plugin": "^0.4.2",
     "lodash": "^4.11.1",
     "memory-fs": "^0.4.1",
@@ -68,7 +70,6 @@
     "raw-loader": "^0.5.1",
     "pug-html-loader": "^1.1.4",
     "resolve": "^1.1.7",
-    "rimraf": "^2.5.3",
     "rsvp": "^3.0.17",
     "rxjs": "^5.0.1",
     "sass-loader": "^6.0.3",
@@ -84,11 +85,11 @@
     "typescript": ">=2.0.0 <2.4.0",
     "url-loader": "^0.5.7",
     "walk-sync": "^0.3.1",
-    "webpack": "~2.4.0",
-    "webpack-dev-middleware": "^1.10.2",
-    "webpack-dev-server": "~2.4.5",
+    "webpack": "~3.3.0",
+    "webpack-dev-middleware": "^1.11.0",
+    "webpack-dev-server": "~2.5.1",
     "webpack-merge": "^2.4.0",
-    "zone.js": "^0.8.4"
+    "zone.js": "^0.8.14"
   },
   "optionalDependencies": {
     "node-sass": "^4.3.0"
