@@ -1,29 +1,13 @@
 import * as fs from 'fs-extra';
-import { expect } from 'chai';
 import * as path from 'path';
-
-const ng = require('../helpers/ng');
-const tmp = require('../helpers/tmp');
-const SilentError = require('silent-error');
+import { ng, setupProject } from '../helpers';
 
 const root = process.cwd();
 
 describe('Acceptance: ng generate guard', function () {
-  beforeEach(() => {
-    return tmp.setup('./tmp').then(() => {
-      process.chdir('./tmp');
-    }).then(() => {
-      return ng(['new', 'foo', '--skip-install']);
-    });
-  });
+  setupProject();
 
-  afterEach(() => {
-    this.timeout(10000);
-
-    return tmp.teardown('./tmp');
-  });
-
-  it('ng generate guard my-guard', () => {
+  it('ng generate guard my-guard', (done) => {
     const appRoot = path.join(root, 'tmp/foo');
     const testPath = path.join(appRoot, 'src/app/my-guard.guard.ts');
     const testSpecPath = path.join(appRoot, 'src/app/my-guard.guard.spec.ts');
@@ -31,17 +15,18 @@ describe('Acceptance: ng generate guard', function () {
 
     return ng(['generate', 'guard', 'my-guard'])
       .then(() => {
-        expect(fs.pathExistsSync(testPath)).to.equal(true);
-        expect(fs.pathExistsSync(testSpecPath)).to.equal(true);
+        expect(fs.pathExistsSync(testPath)).toBe(true);
+        expect(fs.pathExistsSync(testSpecPath)).toBe(true);
       })
       .then(() => fs.readFile(appModulePath, 'utf-8'))
       .then((content: string) => {
-        expect(content).not.to.matches(/import.*MyGuardGuard.*from '.\/my-guard.guard';/);
-        expect(content).not.to.matches(/providers:\s*\[MyGuardGuard\]/m);
-      });
+        expect(content).not.toMatch(/import.*MyGuardGuard.*from '.\/my-guard.guard';/);
+        expect(content).not.toMatch(/providers:\s*\[MyGuardGuard\]/m);
+      })
+      .then(done, done.fail);
   });
 
-  it('ng generate guard my-guard --no-spec', () => {
+  it('ng generate guard my-guard --no-spec', (done) => {
     const appRoot = path.join(root, 'tmp/foo');
     const testPath = path.join(appRoot, 'src/app/my-guard.guard.ts');
     const testSpecPath = path.join(appRoot, 'src/app/my-guard.guard.spec.ts');
@@ -49,32 +34,35 @@ describe('Acceptance: ng generate guard', function () {
 
     return ng(['generate', 'guard', 'my-guard', '--no-spec'])
       .then(() => {
-        expect(fs.pathExistsSync(testPath)).to.equal(true);
-        expect(fs.pathExistsSync(testSpecPath)).to.equal(false);
+        expect(fs.pathExistsSync(testPath)).toBe(true);
+        expect(fs.pathExistsSync(testSpecPath)).toBe(false);
       })
       .then(() => fs.readFile(appModulePath, 'utf-8'))
       .then((content: string) => {
-        expect(content).not.to.matches(/import.*MyGuardGuard.*from '.\/my-guard.guard';/);
-        expect(content).not.to.matches(/providers:\s*\[MyGuardGuard\]/m);
-      });
+        expect(content).not.toMatch(/import.*MyGuardGuard.*from '.\/my-guard.guard';/);
+        expect(content).not.toMatch(/providers:\s*\[MyGuardGuard\]/m);
+      })
+      .then(done, done.fail);
   });
 
-  it('ng generate guard test' + path.sep + 'my-guard', () => {
+  it('ng generate guard test' + path.sep + 'my-guard', (done) => {
     fs.mkdirsSync(path.join(root, 'tmp', 'foo', 'src', 'app', 'test'));
     return ng(['generate', 'guard', 'test' + path.sep + 'my-guard']).then(() => {
       const testPath = path.join(root, 'tmp', 'foo', 'src', 'app', 'test', 'my-guard.guard.ts');
-      expect(fs.pathExistsSync(testPath)).to.equal(true);
-    });
+      expect(fs.pathExistsSync(testPath)).toBe(true);
+    })
+    .then(done, done.fail);
   });
 
-  it('ng generate guard test' + path.sep + '..' + path.sep + 'my-guard', () => {
+  it('ng generate guard test' + path.sep + '..' + path.sep + 'my-guard', (done) => {
     return ng(['generate', 'guard', 'test' + path.sep + '..' + path.sep + 'my-guard']).then(() => {
       const testPath = path.join(root, 'tmp', 'foo', 'src', 'app', 'my-guard.guard.ts');
-      expect(fs.pathExistsSync(testPath)).to.equal(true);
-    });
+      expect(fs.pathExistsSync(testPath)).toBe(true);
+    })
+    .then(done, done.fail);
   });
 
-  it('ng generate guard my-guard from a child dir', () => {
+  it('ng generate guard my-guard from a child dir', (done) => {
     fs.mkdirsSync(path.join(root, 'tmp', 'foo', 'src', 'app', '1'));
     return new Promise(function (resolve) {
       process.chdir('./src');
@@ -88,11 +76,12 @@ describe('Acceptance: ng generate guard', function () {
       })
       .then(() => {
         const testPath = path.join(root, 'tmp', 'foo', 'src', 'app', '1', 'my-guard.guard.ts');
-        expect(fs.pathExistsSync(testPath)).to.equal(true);
-      });
+        expect(fs.pathExistsSync(testPath)).toBe(true);
+      })
+      .then(done, done.fail);
   });
 
-  it('ng generate guard child-dir' + path.sep + 'my-guard from a child dir', () => {
+  it('ng generate guard child-dir' + path.sep + 'my-guard from a child dir', (done) => {
     fs.mkdirsSync(path.join(root, 'tmp', 'foo', 'src', 'app', '1', 'child-dir'));
     return new Promise(function (resolve) {
       process.chdir('./src');
@@ -107,12 +96,13 @@ describe('Acceptance: ng generate guard', function () {
       .then(() => {
         const testPath = path.join(
           root, 'tmp', 'foo', 'src', 'app', '1', 'child-dir', 'my-guard.guard.ts');
-        expect(fs.pathExistsSync(testPath)).to.equal(true);
-      });
+        expect(fs.pathExistsSync(testPath)).toBe(true);
+      })
+      .then(done, done.fail);
   });
 
   it('ng generate guard child-dir' + path.sep + '..' + path.sep + 'my-guard from a child dir',
-    () => {
+    (done) => {
       fs.mkdirsSync(path.join(root, 'tmp', 'foo', 'src', 'app', '1'));
       return new Promise(function (resolve) {
         process.chdir('./src');
@@ -128,13 +118,14 @@ describe('Acceptance: ng generate guard', function () {
         .then(() => {
           const testPath =
             path.join(root, 'tmp', 'foo', 'src', 'app', '1', 'my-guard.guard.ts');
-          expect(fs.pathExistsSync(testPath)).to.equal(true);
-        });
+          expect(fs.pathExistsSync(testPath)).toBe(true);
+        })
+        .then(done, done.fail);
     });
 
   it('ng generate guard ' + path.sep + 'my-guard from a child dir, gens under ' +
     path.join('src', 'app'),
-    () => {
+    (done) => {
       fs.mkdirsSync(path.join(root, 'tmp', 'foo', 'src', 'app', '1'));
       return new Promise(function (resolve) {
         process.chdir('./src');
@@ -148,29 +139,33 @@ describe('Acceptance: ng generate guard', function () {
         })
         .then(() => {
           const testPath = path.join(root, 'tmp', 'foo', 'src', 'app', 'my-guard.guard.ts');
-          expect(fs.pathExistsSync(testPath)).to.equal(true);
-        });
+          expect(fs.pathExistsSync(testPath)).toBe(true);
+        })
+        .then(done, done.fail);
     });
 
-  it('ng generate guard ..' + path.sep + 'my-guard from root dir will fail', () => {
-    return ng(['generate', 'guard', '..' + path.sep + 'my-guard']).then(() => {
-      throw new SilentError(`ng generate guard ..${path.sep}my-guard from root dir should fail.`);
-    }, (err: string) => {
-      // tslint:disable-next-line:max-line-length
-      expect(err).to.equal(`Invalid path: "..${path.sep}my-guard" cannot be above the "src${path.sep}app" directory`);
-    });
+  it('ng generate guard ..' + path.sep + 'my-guard from root dir will fail', (done) => {
+    return ng(['generate', 'guard', '..' + path.sep + 'my-guard'])
+      .then(() => done.fail())
+      .catch(err => {
+        // tslint:disable-next-line:max-line-length
+        expect(err).toBe(`Invalid path: "..${path.sep}my-guard" cannot be above the "src${path.sep}app" directory`);
+      })
+      .then(done, done.fail);
   });
 
-  it('should error out when given an incorrect module path', () => {
+  it('should error out when given an incorrect module path', (done) => {
     return Promise.resolve()
       .then(() => ng(['generate', 'guard', 'baz', '--module', 'foo']))
+      .then(() => done.fail())
       .catch((error) => {
-        expect(error).to.equal('Specified module does not exist');
-      });
+        expect(error).toBe('Specified module does not exist');
+      })
+      .then(done, done.fail);
   });
 
   describe('should import and add to provider list', () => {
-    it('when given a root level module with module.ts suffix', () => {
+    it('when given a root level module with module.ts suffix', (done) => {
       const appRoot = path.join(root, 'tmp/foo');
       const modulePath = path.join(appRoot, 'src/app/app.module.ts');
 
@@ -178,12 +173,13 @@ describe('Acceptance: ng generate guard', function () {
         .then(() => ng(['generate', 'guard', 'baz', '--module', 'app.module.ts']))
         .then(() => fs.readFile(modulePath, 'utf-8'))
         .then(content => {
-          expect(content).to.matches(/import.*BazGuard.*from '.\/baz.guard';/);
-          expect(content).to.matches(/providers:\s*\[BazGuard\]/m);
-        });
+          expect(content).toMatch(/import.*BazGuard.*from '.\/baz.guard';/);
+          expect(content).toMatch(/providers:\s*\[BazGuard\]/m);
+        })
+        .then(done, done.fail);
     });
 
-    it('when given a root level module with missing module.ts suffix', () => {
+    it('when given a root level module with missing module.ts suffix', (done) => {
       const appRoot = path.join(root, 'tmp/foo');
       const modulePath = path.join(appRoot, 'src/app/app.module.ts');
 
@@ -191,12 +187,13 @@ describe('Acceptance: ng generate guard', function () {
         .then(() => ng(['generate', 'guard', 'baz', '--module', 'app']))
         .then(() => fs.readFile(modulePath, 'utf-8'))
         .then(content => {
-          expect(content).to.matches(/import.*BazGuard.*from '.\/baz.guard';/);
-          expect(content).to.matches(/providers:\s*\[BazGuard\]/m);
-        });
+          expect(content).toMatch(/import.*BazGuard.*from '.\/baz.guard';/);
+          expect(content).toMatch(/providers:\s*\[BazGuard\]/m);
+        })
+        .then(done, done.fail);
     });
 
-    it('when given a submodule with module.ts suffix', () => {
+    it('when given a submodule with module.ts suffix', (done) => {
       const appRoot = path.join(root, 'tmp/foo');
       const modulePath = path.join(appRoot, 'src/app/foo/foo.module.ts');
 
@@ -205,12 +202,13 @@ describe('Acceptance: ng generate guard', function () {
         .then(() => ng(['generate', 'guard', 'baz', '--module', path.join('foo', 'foo.module.ts')]))
         .then(() => fs.readFile(modulePath, 'utf-8'))
         .then(content => {
-          expect(content).to.matches(/import.*BazGuard.*from '..\/baz.guard';/);
-          expect(content).to.matches(/providers:\s*\[BazGuard\]/m);
-        });
+          expect(content).toMatch(/import.*BazGuard.*from '..\/baz.guard';/);
+          expect(content).toMatch(/providers:\s*\[BazGuard\]/m);
+        })
+        .then(done, done.fail);
     });
 
-    it('when given a submodule with missing module.ts suffix', () => {
+    it('when given a submodule with missing module.ts suffix', (done) => {
       const appRoot = path.join(root, 'tmp/foo');
       const modulePath = path.join(appRoot, 'src/app/foo/foo.module.ts');
 
@@ -219,12 +217,13 @@ describe('Acceptance: ng generate guard', function () {
         .then(() => ng(['generate', 'guard', 'baz', '--module', path.join('foo', 'foo')]))
         .then(() => fs.readFile(modulePath, 'utf-8'))
         .then(content => {
-          expect(content).to.matches(/import.*BazGuard.*from '..\/baz.guard';/);
-          expect(content).to.matches(/providers:\s*\[BazGuard\]/m);
-        });
+          expect(content).toMatch(/import.*BazGuard.*from '..\/baz.guard';/);
+          expect(content).toMatch(/providers:\s*\[BazGuard\]/m);
+        })
+        .then(done, done.fail);
     });
 
-    it('when given a submodule folder', () => {
+    it('when given a submodule folder', (done) => {
       const appRoot = path.join(root, 'tmp/foo');
       const modulePath = path.join(appRoot, 'src/app/foo/foo.module.ts');
 
@@ -233,12 +232,13 @@ describe('Acceptance: ng generate guard', function () {
         .then(() => ng(['generate', 'guard', 'baz', '--module', 'foo']))
         .then(() => fs.readFile(modulePath, 'utf-8'))
         .then(content => {
-          expect(content).to.matches(/import.*BazGuard.*from '..\/baz.guard';/);
-          expect(content).to.matches(/providers:\s*\[BazGuard\]/m);
-        });
+          expect(content).toMatch(/import.*BazGuard.*from '..\/baz.guard';/);
+          expect(content).toMatch(/providers:\s*\[BazGuard\]/m);
+        })
+        .then(done, done.fail);
     });
 
-    it('when given deep submodule folder with missing module.ts suffix', () => {
+    it('when given deep submodule folder with missing module.ts suffix', (done) => {
       const appRoot = path.join(root, 'tmp/foo');
       const modulePath = path.join(appRoot, 'src/app/foo/bar/bar.module.ts');
 
@@ -248,9 +248,10 @@ describe('Acceptance: ng generate guard', function () {
         .then(() => ng(['generate', 'guard', 'baz', '--module', path.join('foo', 'bar')]))
         .then(() => fs.readFile(modulePath, 'utf-8'))
         .then(content => {
-          expect(content).to.matches(/import.*BazGuard.*from '..\/..\/baz.guard';/);
-          expect(content).to.matches(/providers:\s*\[BazGuard\]/m);
-        });
+          expect(content).toMatch(/import.*BazGuard.*from '..\/..\/baz.guard';/);
+          expect(content).toMatch(/providers:\s*\[BazGuard\]/m);
+        })
+        .then(done, done.fail);
     });
   });
 });
