@@ -27,7 +27,8 @@ const packages =
         packageJson: path.join(pkg.root, 'package.json'),
         root: pkg.root,
         relative: path.relative(path.dirname(__dirname), pkg.root),
-        main: path.resolve(pkg.root, 'src/index.ts')
+        main: path.resolve(pkg.root, 'src/index.ts'),
+        tar: path.join(distRoot, pkg.name.replace('/', '_') + '.tgz')
       };
       return packages;
     }, {});
