@@ -31,6 +31,6 @@ describe('AppComponent', () => {
     const fixture = TestBed.createComponent(AppComponent);
     fixture.detectChanges();
     const compiled = fixture.debugElement.nativeElement;
-    expect(compiled.querySelector('h1').textContent).toContain('Welcome to <%= prefix %>!!');
+    expect(compiled.querySelector('h1').textContent).toContain('Welcome to <%= prefix %>!');
   }));
 });
