@@ -7,7 +7,7 @@ export interface DocOptions {
 
 const DocCommand = Command.extend({
   name: 'doc',
-  description: 'Opens the official Angular documentation for a given keyword.',
+  description: 'Opens the official Angular API documentation for a given keyword.',
   works: 'everywhere',
   availableOptions: [
     {
@@ -15,7 +15,7 @@ const DocCommand = Command.extend({
       aliases: ['s'],
       type: Boolean,
       default: false,
-      description: 'Search docs instead of api.'
+      description: 'Search whole angular.io instead of just api.'
     }
   ],
 
