@@ -42,6 +42,24 @@ export function getBrowserConfig(wco: WebpackConfigOptions) {
     }));
   }
 
+  if (buildOptions.sourcemaps) {
+    extraPlugins.push(new webpack.SourceMapDevToolPlugin({
+      filename: '[file].map[query]',
+      moduleFilenameTemplate: '[resource-path]',
+      fallbackModuleFilenameTemplate: '[resource-path]?[hash]',
+      sourceRoot: 'webpack:///'
+    }));
+  }
+
+  if (buildOptions.commonChunk) {
+    extraPlugins.push(new webpack.optimize.CommonsChunkPlugin({
+      name: 'main',
+      async: 'common',
+      children: true,
+      minChunks: 2
+    }));
+  }
+
   return {
     plugins: [
       new HtmlWebpackPlugin({
@@ -58,11 +76,6 @@ export function getBrowserConfig(wco: WebpackConfigOptions) {
       }),
       new BaseHrefWebpackPlugin({
         baseHref: buildOptions.baseHref
-      }),
-      new webpack.optimize.CommonsChunkPlugin({
-        async: 'common',
-        children: true,
-        minChunks: 2
       }),
       new webpack.optimize.CommonsChunkPlugin({
         minChunks: Infinity,
