@@ -6,5 +6,7 @@ module.exports = {
   GlobCopyWebpackPlugin: require('../plugins/glob-copy-webpack-plugin').GlobCopyWebpackPlugin,
   SuppressExtractedTextChunksWebpackPlugin:
     require('../plugins/suppress-entry-chunks-webpack-plugin')
-      .SuppressExtractedTextChunksWebpackPlugin
+      .SuppressExtractedTextChunksWebpackPlugin,
+  NamedLazyChunksWebpackPlugin:
+    require('../plugins/named-lazy-chunks-webpack-plugin').NamedLazyChunksWebpackPlugin
 };
