@@ -3,8 +3,8 @@ const opn = require('opn');
 
 export const DocTask: any = Task.extend({
   run: function(keyword: string, search: boolean) {
-    const searchUrl = search ? `https://angular.io/search/#stq=${keyword}&stp=1` :
-     `https://angular.io/docs/ts/latest/api/#!?query=${keyword}`;
+    const searchUrl = search ? `https://www.google.com/search?q=site%3Aangular.io+${keyword}` :
+     `https://angular.io/api?query=${keyword}`;
 
     return opn(searchUrl, { wait: false });
   }
