@@ -2,8 +2,6 @@ import mockFs = require('mock-fs');
 import * as ts from 'typescript';
 import * as path from 'path';
 import * as dependentFilesUtils from '@angular/cli/utilities/get-dependent-files';
-
-import { expect } from 'chai';
 import { ModuleResolver } from '@angular/cli/utilities/module-resolver';
 
 describe('ModuleResolver', () => {
@@ -59,7 +57,7 @@ describe('ModuleResolver', () => {
     let fooBarFile = path.join(rootPath, 'foo-baz/qux/quux/foobar/foobar.component.ts');
     let fooQuxFile = path.join(rootPath, 'foo-baz/qux/quux/fooqux/fooqux.component.ts');
 
-    it('when there is no index.ts in oldPath', () => {
+    it('when there is no index.ts in oldPath', (done) => {
       let oldFilePath = path.join(rootPath, 'bar/baz/baz.component.ts');
       let newFilePath = path.join(rootPath, 'foo');
       let resolver = new ModuleResolver(oldFilePath, newFilePath, rootPath);
@@ -69,13 +67,13 @@ describe('ModuleResolver', () => {
         .then((tsFileBar: ts.SourceFile) => {
           let contentsBar = dependentFilesUtils.getImportClauses(tsFileBar);
           let bazExpectedContent = path.normalize('../foo/baz.component');
-          expect(contentsBar[0].specifierText).to.equal(bazExpectedContent);
+          expect(contentsBar[0].specifierText).toBe(bazExpectedContent);
         })
         .then(() => dependentFilesUtils.createTsSourceFile(fooFile))
         .then((tsFileFoo: ts.SourceFile) => {
           let contentsFoo = dependentFilesUtils.getImportClauses(tsFileFoo);
           let bazExpectedContent = './baz.component'.replace('/', path.sep);
-          expect(contentsFoo[0].specifierText).to.equal(bazExpectedContent);
+          expect(contentsFoo[0].specifierText).toBe(bazExpectedContent);
         })
         .then(() => resolver.resolveOwnImports())
         .then((changes) => resolver.applySortedChangePromise(changes))
@@ -84,11 +82,12 @@ describe('ModuleResolver', () => {
           let contentsBaz = dependentFilesUtils.getImportClauses(tsFileBaz);
           let barExpectedContent = path.normalize('../bar/bar.component');
           let fooBarExpectedContent = path.normalize('../foo-baz/qux/quux/foobar/foobar.component');
-          expect(contentsBaz[0].specifierText).to.equal(barExpectedContent);
-          expect(contentsBaz[1].specifierText).to.equal(fooBarExpectedContent);
-        });
+          expect(contentsBaz[0].specifierText).toBe(barExpectedContent);
+          expect(contentsBaz[1].specifierText).toBe(fooBarExpectedContent);
+        })
+        .then(done, done.fail);
     });
-    it('when no files are importing the given file', () => {
+    it('when no files are importing the given file', (done) => {
       let oldFilePath = path.join(rootPath, 'foo-baz/foo-baz.component.ts');
       let newFilePath = path.join(rootPath, 'bar');
       let resolver = new ModuleResolver(oldFilePath, newFilePath, rootPath);
@@ -100,10 +99,11 @@ describe('ModuleResolver', () => {
         .then((tsFile: ts.SourceFile) => {
           let contents = dependentFilesUtils.getImportClauses(tsFile);
           let fooExpectedContent = path.normalize('../foo/foo.component');
-          expect(contents[0].specifierText).to.equal(fooExpectedContent);
-        });
+          expect(contents[0].specifierText).toBe(fooExpectedContent);
+        })
+        .then(done, done.fail);
     });
-    it('when oldPath and newPath both do not have index.ts', () => {
+    it('when oldPath and newPath both do not have index.ts', (done) => {
       let oldFilePath = path.join(rootPath, 'bar/baz/baz.component.ts');
       let newFilePath = path.join(rootPath, 'foo-baz');
       let resolver = new ModuleResolver(oldFilePath, newFilePath, rootPath);
@@ -113,13 +113,13 @@ describe('ModuleResolver', () => {
         .then((tsFileBar: ts.SourceFile) => {
           let contentsBar = dependentFilesUtils.getImportClauses(tsFileBar);
           let bazExpectedContent = path.normalize('../foo-baz/baz.component');
-          expect(contentsBar[0].specifierText).to.equal(bazExpectedContent);
+          expect(contentsBar[0].specifierText).toBe(bazExpectedContent);
         })
         .then(() => dependentFilesUtils.createTsSourceFile(fooFile))
         .then((tsFileFoo: ts.SourceFile) => {
           let contentsFoo = dependentFilesUtils.getImportClauses(tsFileFoo);
           let bazExpectedContent = path.normalize('../foo-baz/baz.component');
-          expect(contentsFoo[0].specifierText).to.equal(bazExpectedContent);
+          expect(contentsFoo[0].specifierText).toBe(bazExpectedContent);
         })
         .then(() => resolver.resolveOwnImports())
         .then((changes) => resolver.applySortedChangePromise(changes))
@@ -129,11 +129,12 @@ describe('ModuleResolver', () => {
           let barExpectedContent = path.normalize('../bar/bar.component');
           let fooBarExpectedContent =
             `.${path.sep}qux${path.sep}quux${path.sep}foobar${path.sep}foobar.component`;
-          expect(contentsBaz[0].specifierText).to.equal(barExpectedContent);
-          expect(contentsBaz[1].specifierText).to.equal(fooBarExpectedContent);
-        });
+          expect(contentsBaz[0].specifierText).toBe(barExpectedContent);
+          expect(contentsBaz[1].specifierText).toBe(fooBarExpectedContent);
+        })
+        .then(done, done.fail);
     });
-    it('when there are multiple spaces between symbols and specifier', () => {
+    it('when there are multiple spaces between symbols and specifier', (done) => {
       let oldFilePath = path.join(rootPath, 'foo-baz/qux/quux/foobar/foobar.component.ts');
       let newFilePath = path.join(rootPath, 'foo');
       let resolver = new ModuleResolver(oldFilePath, newFilePath, rootPath);
@@ -143,13 +144,13 @@ describe('ModuleResolver', () => {
         .then((tsFileFooQux: ts.SourceFile) => {
           let contentsFooQux = dependentFilesUtils.getImportClauses(tsFileFooQux);
           let fooQuxExpectedContent = path.normalize('../../../../foo/foobar.component');
-          expect(contentsFooQux[0].specifierText).to.equal(fooQuxExpectedContent);
+          expect(contentsFooQux[0].specifierText).toBe(fooQuxExpectedContent);
         })
         .then(() => dependentFilesUtils.createTsSourceFile(bazFile))
         .then((tsFileBaz: ts.SourceFile) => {
           let contentsBaz = dependentFilesUtils.getImportClauses(tsFileBaz);
           let bazExpectedContent = path.normalize('../../foo/foobar.component');
-          expect(contentsBaz[1].specifierText).to.equal(bazExpectedContent);
+          expect(contentsBaz[1].specifierText).toBe(bazExpectedContent);
         })
         .then(() => resolver.resolveOwnImports())
         .then((changes) => resolver.applySortedChangePromise(changes))
@@ -158,9 +159,10 @@ describe('ModuleResolver', () => {
           let contentsFooBar = dependentFilesUtils.getImportClauses(tsFileFooBar);
           let fooExpectedContent = `.${path.sep}foo.component`;
           let fooQuxExpectedContent = path.normalize('../foo-baz/qux/quux/fooqux.component');
-          expect(contentsFooBar[0].specifierText).to.equal(fooExpectedContent);
-          expect(contentsFooBar[1].specifierText).to.equal(fooQuxExpectedContent);
-        });
+          expect(contentsFooBar[0].specifierText).toBe(fooExpectedContent);
+          expect(contentsFooBar[1].specifierText).toBe(fooQuxExpectedContent);
+        })
+        .then(done, done.fail);
     });
   });
 });
