@@ -1,5 +1,5 @@
 import {readFile, writeFile} from './fs';
-import {silentExecAndWaitForOutputToMatch, silentNpm, ng} from './process';
+import {execAndWaitForOutputToMatch, silentNpm, ng} from './process';
 import {getGlobalVariable} from './env';
 
 const packages = require('../../../lib/packages').packages;
@@ -25,8 +25,8 @@ export function updateTsConfig(fn: (json: any) => any | void) {
 
 
 export function ngServe(...args: string[]) {
-  return silentExecAndWaitForOutputToMatch('ng',
-    ['serve', '--no-progress', ...args],
+  return execAndWaitForOutputToMatch('ng',
+    ['serve', ...args],
     /webpack: bundle is now VALID|webpack: Compiled successfully./);
 }
 
@@ -93,6 +93,8 @@ export function useNg2() {
     },
     'devDependencies': {
       '@angular/compiler-cli': '^2.4.0',
+      '@types/jasmine': '~2.2.0',
+      '@types/jasminewd2': undefined,
       'typescript': '~2.0.0'
     }
   };
@@ -106,7 +108,7 @@ export function useNg2() {
       'experimentalDecorators': true,
       'target': 'es5',
       'lib': [
-        'es2016',
+        'es2017',
         'dom'
       ],
       'outDir': '../out-tsc/app',
@@ -128,7 +130,7 @@ export function useNg2() {
       'emitDecoratorMetadata': true,
       'experimentalDecorators': true,
       'lib': [
-        'es2016',
+        'es2017',
         'dom'
       ],
       'outDir': '../out-tsc/spec',
@@ -157,7 +159,7 @@ export function useNg2() {
       'emitDecoratorMetadata': true,
       'experimentalDecorators': true,
       'lib': [
-        'es2016'
+        'es2017'
       ],
       'outDir': '../out-tsc/e2e',
       'module': 'commonjs',
